% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict-data.R
\docType{data}
\name{dict}
\alias{dict}
\title{Dict data}
\format{A dict dataset.}
\usage{
data(dict)
}
\description{
Data from a dict.
}
\examples{
data(dict)
d <- length(dict)

}
\keyword{datasets}
