% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cooc.R
\name{testCooc}
\alias{testCooc}
\title{testCooc}
\usage{
testCooc(data = "")
}
\arguments{
\item{data}{a character string.}
}
\description{
This function runs a test.
}
\examples{
testCooc()
}
\keyword{cooc}
